Split amp-stream-gallery
