+++ extensions/amp-stream-gallery/1.0/component.js
